Extract portfolio auth strategy
